pyproject.toml


[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.3"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
